Bail out if git clone failed
If the user specifies an unknown branch or tag, the git clone will fail.
Previously, the clone would work but checking out the branch would fail,
allowing the generator to continue using the master branch or
`lockedIgniteVersion` tag. Now, however, the clone will fail if the
branch is unknown, so bail out immediately.