during startup, we now have a check if everything is loaded already from storage.
for example, before, if a user was logged in, the log in button would show up for a split second while the async storage loaded, and when it was finally loaded, it would swap for a log out button. now, we can have a check if the storage is loaded, so we can display an animation or a spinner, and then immediately show the log out button. no more flashing from log in -> log out.