Remove flexible version specifiers from our dependencies
This ensures that apps we generate will use the same immediate
dependencies as what we used when we test a new version of Ignite.

To do this, I removed the "^" and "~" from versions in package.json;
for a few dependencies, I changed the version to match what yarn.lock
tells me we used when we tested 6.1.1.

After doing this, I re-ran `yarn install`, which updated yarn.lock,
but the only changes there are to lines that record the requirements,
not the actual versions resolved - which tells me I got all the
versions right.

our postinstall script (which yarn install runs) did change
Podfile.lock - looks like a bunch of react-native ; I'm not sure why,
but I'm committing it anyway.