Workaround for 0.74.5 upgrade
Without this patch, the JS bundle fails to load with:

```
iOS Bundling failed 3255ms node_modules/expo/AppEntry.js (1892 modules)
Unable to resolve "react-native/Libraries/Utilities/DevLoadingView" from "node_modules/@expo/metro-runtime/build/LoadingView.native.js"
```

This is in a conditional in LoadingView.native.js that should load the
correct one, and does, so this appears to be an issue with conditional
requires. I'm not sure how to fix that underlying issue just yet, so
patching for now.