fix(tests): fix i18n tests
The tests were failing on #2924 but i thought it was just the CI being flaky… surely changing one word in a comment wouldn’t cause the entire test suite to hang!

Boy was I wrong. Turns out that the i18n tests actually greps the filesystem looking for instances of the translate() function, extracts their keys, and then compares those keys to the list inside the translation files.

Well, this grep command doesn’t take commented code into account so it sees these blocks of code as equally valid:

translate(‘key1’)

// translate(‘key2’)

/**
  * translate(‘key3’)
  */

So the test will look for [key1, key2, key3] and make sure the translations exist.

To solve this, I added the key `”hello”` to the EXCEPTIONS list so that this comment will no longer cause the test suite to hang and eventually fail.